ai-service/requirements.txt
# ===========================================
# NutriVision AI - Python AI Microservice
# ===========================================

# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# AI/ML Libraries
torch==2.1.2
torchvision==0.16.2
ultralytics==8.1.0
opencv-python-headless==4.9.0.80
Pillow==10.2.0
numpy==1.26.3
scikit-learn==1.4.0

# GenAI Integration
openai==1.8.0
anthropic==0.8.1
langchain==0.1.0
tiktoken==0.5.2
transformers==4.36.2

# Database Clients
asyncpg==0.29.0
motor==3.3.2
redis==5.0.1
sqlalchemy==2.0.25

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1

# Monitoring & Logging
structlog==24.1.0
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
